Remove legacy stickyNote API
