+++ designer/client/src/components/graph/types.ts
@@ -69,7 +69,6 @@ export enum Events {
     CELL_MOVED = "cellCustom:moved",
     CELL_RESIZED = "cellCustom:resized",
     CELL_CONTENT_UPDATED = "cellCustom:contentUpdated",
-    CELL_DELETED = "cellCustom:deleted",
     BLANK_POINTERCLICK = "blank:pointerclick",
     BLANK_POINTERDOWN = "blank:pointerdown",
     BLANK_POINTERUP = "blank:pointerup",
