+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/SettingsResources.scala
@@ -6,8 +6,8 @@ import io.circe.{Decoder, Encoder}
 import io.circe.generic.JsonCodec
 import org.apache.pekko.http.scaladsl.server.{Directives, Route}
 import pl.touk.nussknacker.engine.api.CirceUtil.codecs._
-import pl.touk.nussknacker.ui.api.description.stickynotes.Dtos.StickyNotesSettings
-import pl.touk.nussknacker.ui.config.{DesignerConfig, UsageStatisticsReportsConfig}
+import pl.touk.nussknacker.ui.api.description.stickynotes.StickyNotesSettings
+import pl.touk.nussknacker.ui.config.{FeatureTogglesConfig, UsageStatisticsReportsConfig}
 import pl.touk.nussknacker.ui.statistics.{Fingerprint, FingerprintService}
 
 import java.net.URL
