+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/description/stickynotes/StickyNotesSettings.scala
@@ -1,0 +1,13 @@
+package pl.touk.nussknacker.ui.api.description.stickynotes
+
+import io.circe.generic.JsonCodec
+
+@JsonCodec case class StickyNotesSettings(
+    maxContentLength: Int,
+    maxNotesCount: Option[Int] = None,
+    enabled: Boolean
+)
+
+object StickyNotesSettings {
+  val default: StickyNotesSettings = StickyNotesSettings(maxContentLength = 5000, enabled = true)
+}
