+++ designer/server/src/test/scala/pl/touk/nussknacker/test/utils/domain/ScenarioHelper.scala
@@ -10,19 +10,13 @@ import pl.touk.nussknacker.engine.management.FlinkStreamingPropertiesConfig
 import pl.touk.nussknacker.test.PatientScalaFutures
 import pl.touk.nussknacker.test.config.WithSimplifiedDesignerConfig.TestProcessingType.Streaming
 import pl.touk.nussknacker.test.mock.TestAdditionalUIConfigProvider
-import pl.touk.nussknacker.ui.api.description.stickynotes.Dtos.{StickyNoteAddRequest, StickyNoteCorrelationId}
 import pl.touk.nussknacker.ui.db.DbRef
 import pl.touk.nussknacker.ui.definition.ScenarioPropertiesConfigFinalizer
 import pl.touk.nussknacker.ui.process.NewProcessPreparer
 import pl.touk.nussknacker.ui.process.processingtype.ValueWithRestriction
 import pl.touk.nussknacker.ui.process.repository._
-import pl.touk.nussknacker.ui.process.repository.ProcessRepository.{
-  CreateProcessAction,
-  ProcessUpdated,
-  UpdateProcessAction
-}
+import pl.touk.nussknacker.ui.process.repository.ProcessRepository.CreateProcessAction
 import pl.touk.nussknacker.ui.process.repository.activities.DbScenarioActivityRepository
-import pl.touk.nussknacker.ui.process.repository.stickynotes.{DbStickyNotesRepository, StickyNotesRepository}
 import pl.touk.nussknacker.ui.security.api.{LoggedUser, RealLoggedUser}
 import slick.dbio.DBIOAction
 
@@ -41,7 +35,6 @@ private[test] class ScenarioHelper(dbRef: DbRef, clock: Clock, designerConfig: C
   private val actionRepository: ScenarioActionRepository = DbScenarioActionRepository.create(dbRef)
 
   private val scenarioLabelsRepository: ScenarioLabelsRepository = new ScenarioLabelsRepository(dbRef)
-  private val stickyNotesRepository: StickyNotesRepository       = DbStickyNotesRepository.create(dbRef, clock)
 
   private val writeScenarioRepository: DBProcessRepository = new DBProcessRepository(
     dbRef,
@@ -76,20 +69,6 @@ private[test] class ScenarioHelper(dbRef: DbRef, clock: Clock, designerConfig: C
       isFragment: Boolean
   ): ProcessId = {
     saveAndGetId(scenario, category, isFragment).futureValue
-  }
-
-  def updateScenario(
-      scenarioName: ProcessName,
-      newScenario: CanonicalProcess
-  ): ProcessUpdated = {
-    updateAndGetScenarioVersions(scenarioName, newScenario).futureValue
-  }
-
-  def addStickyNote(
-      scenarioName: ProcessName,
-      request: StickyNoteAddRequest
-  ): StickyNoteCorrelationId = {
-    addStickyNoteForScenario(scenarioName, request).futureValue
   }
 
   def createDeployedExampleScenario(scenarioName: ProcessName, category: String, isFragment: Boolean): ProcessId = {
@@ -203,44 +182,6 @@ private[test] class ScenarioHelper(dbRef: DbRef, clock: Clock, designerConfig: C
       _  <- dbioRunner.runInSerializableTransactionWithRetry(writeScenarioRepository.saveNewProcess(action))
       id <- futureFetchingScenarioRepository.fetchProcessId(scenarioName).map(_.get)
     } yield id
-  }
-
-  private def updateAndGetScenarioVersions(
-      scenarioName: ProcessName,
-      newScenario: CanonicalProcess
-  ): Future[ProcessUpdated] = {
-    for {
-      scenarioId <- futureFetchingScenarioRepository.fetchProcessId(scenarioName).map(_.get)
-      action = UpdateProcessAction(
-        scenarioId,
-        newScenario,
-        comment = None,
-        labels = List.empty,
-        increaseVersionWhenJsonNotChanged = true,
-        forwardedUserName = None
-      )
-      processUpdated <- dbioRunner.runInTransaction(writeScenarioRepository.updateProcess(action))
-    } yield processUpdated
-  }
-
-  private def addStickyNoteForScenario(
-      scenarioName: ProcessName,
-      request: StickyNoteAddRequest
-  ): Future[StickyNoteCorrelationId] = {
-    for {
-      scenarioId <- futureFetchingScenarioRepository.fetchProcessId(scenarioName).map(_.get)
-      noteCorrelationId <- dbioRunner.runInTransaction(
-        stickyNotesRepository.addStickyNote(
-          request.content,
-          request.layoutData,
-          request.color,
-          request.dimensions,
-          request.targetEdge,
-          scenarioId,
-          request.scenarioVersionId
-        )
-      )
-    } yield noteCorrelationId
   }
 
   private def mapProcessingTypeDataProvider[T](value: T) = {
